refactor(StrikeZoneFactory): strikeZone creation feature move
refactor contents
- added StrikeZoneFactory
- Referee's strikeZone creation feature moved to StrikeZoneFactory
