+++ src/main/java/baseball/game/referee/Referee.java
@@ -1,0 +1,10 @@
+package baseball.game.referee;
+
+public class Referee {
+    private int strikeZone;
+
+    public Referee() {
+        this.strikeZone = StrikeZoneFactory.createStrikeZone();
+    }
+
+}
